Feature: Simultaneous stream playback support
With some tweaks to /etc/asound.conf, it is possible to mix multiple streams
from mplayer (accomplished by multiple instances of mplayer running).

This scenario comes up when you ask AVS a question while a music stream is
playing.  Now we lower the volume of the active music, play the AVS response
at normal volume, then restore the volume of the music.

Note: some hardware may support this natively, but mine does not, and required
help from this page:
https://discuss.mopidy.com/t/setting-up-c-media-usb-headphone-set-in-mixer-mode/1409/3
(I used hw:1,0 for both dmix and dsnoop)

Check the README for more information

Signed-off-by: Mike Erdahl <mike.erdahl@gmail.com>