Fix glitch resuming stream at full volume after AVS response
Was unconditionally playing a streamUrl without checking if we were
already streaming.  This causes a gap in audio while one stream
is stopped, and the other is started.  This would be OK if the streams
were different, but this should not happen if the streams are the same.

It is becoming more clear there needs to be more separation of logic
between the audio files, stream players and ALSA.

Signed-off-by: Mike Erdahl <mike.erdahl@gmail.com>