Feature: restart stalled HTTP audio stream
Sometimes audio streams from the interwebs just die, and it is nice
to detect this, and automatically restart the stream.  MPlayer will
returns None for volume when a stream is not playing, so it is
easy to detect when a stream is dead, but our state variables say
we should be playing a stream.  Make it so!

Signed-off-by: Mike Erdahl <mike.erdahl@gmail.com>